refactor: DTO 구조 변경으로 인한 경로 수정
