+++ src/main/java/leets/weeth/domain/user/application/usecase/UserUseCase.java
@@ -1,9 +1,14 @@
 package leets.weeth.domain.user.application.usecase;
 
+import leets.weeth.domain.user.application.dto.request.UserRequestDto;
+import leets.weeth.domain.user.application.dto.response.UserResponseDto;
+
 import java.util.List;
 import java.util.Map;
 
-import static leets.weeth.domain.user.application.dto.UserDTO.*;
+import static leets.weeth.domain.user.application.dto.request.UserRequestDto.*;
+import static leets.weeth.domain.user.application.dto.response.UserResponseDto.*;
+
 
 public interface UserUseCase {
