+++ src/main/java/leets/weeth/domain/user/application/usecase/UserUseCaseImpl.java
@@ -4,6 +4,8 @@ import jakarta.transaction.Transactional;
 import leets.weeth.domain.attendance.domain.service.AttendanceSaveService;
 import leets.weeth.domain.schedule.domain.entity.Meeting;
 import leets.weeth.domain.schedule.domain.service.MeetingGetService;
+import leets.weeth.domain.user.application.dto.request.UserRequestDto;
+import leets.weeth.domain.user.application.dto.response.UserResponseDto;
 import leets.weeth.domain.user.application.mapper.UserMapper;
 import leets.weeth.domain.user.domain.entity.User;
 import leets.weeth.domain.user.domain.service.UserDeleteService;
@@ -12,6 +14,7 @@ import leets.weeth.domain.user.domain.service.UserSaveService;
 import leets.weeth.domain.user.domain.service.UserUpdateService;
 import leets.weeth.domain.user.application.exception.StudentIdExistsException;
 import leets.weeth.domain.user.application.exception.TelExistsException;
+import leets.weeth.global.auth.jwt.service.JwtService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -22,7 +25,8 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static leets.weeth.domain.user.application.dto.UserDTO.*;
+import static leets.weeth.domain.user.application.dto.request.UserRequestDto.*;
+import static leets.weeth.domain.user.application.dto.response.UserResponseDto.*;
 import static leets.weeth.domain.user.domain.entity.enums.Status.ACTIVE;
 
 @Service
@@ -37,6 +41,8 @@ public class UserUseCaseImpl implements UserUseCase {
     private final PasswordEncoder passwordEncoder;
     private final AttendanceSaveService attendanceSaveService;
     private final MeetingGetService meetingGetService;
+
+    private final JwtService jwtService;
 
     @Override
     public void apply(SignUp dto) {
@@ -95,12 +101,15 @@ public class UserUseCaseImpl implements UserUseCase {
     @Override
     public void leave(Long userId) {
         User user = userGetService.find(userId);
+        // 탈퇴하는 경우 리프레시 토큰 삭제
+        jwtService.delete(user.getEmail());
         userDeleteService.leave(user);
     }
 
     @Override
     public void ban(Long userId) {
         User user = userGetService.find(userId);
+        jwtService.delete(user.getEmail());
         userDeleteService.ban(user);
     }
