+++ src/main/java/leets/weeth/domain/user/domain/service/UserUpdateService.java
@@ -1,12 +1,14 @@
 package leets.weeth.domain.user.domain.service;
 
 import jakarta.transaction.Transactional;
+import leets.weeth.domain.user.application.dto.request.UserRequestDto;
 import leets.weeth.domain.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import static leets.weeth.domain.user.application.dto.UserDTO.Update;
+import static leets.weeth.domain.user.application.dto.request.UserRequestDto.*;
+
 
 @Service
 @Transactional
