+++ src/main/java/leets/weeth/domain/user/presentation/UserController.java
@@ -3,7 +3,8 @@ package leets.weeth.domain.user.presentation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
-import leets.weeth.domain.user.application.dto.UserDTO;
+import leets.weeth.domain.user.application.dto.request.UserRequestDto;
+import leets.weeth.domain.user.application.dto.response.UserResponseDto;
 import leets.weeth.domain.user.application.usecase.UserManageUseCase;
 import leets.weeth.domain.user.application.usecase.UserUseCase;
 import leets.weeth.domain.user.domain.service.UserGetService;
@@ -15,7 +16,6 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Map;
 
-import static leets.weeth.domain.user.application.dto.UserDTO.*;
 
 @Tag(name = "UserController", description = "사용자 관련 컨트롤러")
 @RestController
@@ -28,7 +28,7 @@ public class UserController {
     private final UserGetService userGetService;
 
     @PostMapping("/apply")
-    public CommonResponse<Void> apply(@RequestBody @Valid SignUp dto) {
+    public CommonResponse<Void> apply(@RequestBody @Valid UserRequestDto.SignUp dto) {
         userUseCase.apply(dto);
         return CommonResponse.createSuccess();
     }
@@ -39,17 +39,17 @@ public class UserController {
     }
 
     @GetMapping("/all")
-    public CommonResponse<Map<Integer, List<Response>>> findAll() {
+    public CommonResponse<Map<Integer, List<UserResponseDto.Response>>> findAll() {
         return CommonResponse.createSuccess(userUseCase.findAll());
     }
 
     @GetMapping
-    public CommonResponse<Response> find(@CurrentUser Long userId) {
+    public CommonResponse<UserResponseDto.Response> find(@CurrentUser Long userId) {
         return CommonResponse.createSuccess(userUseCase.find(userId));
     }
 
     @PatchMapping
-    public CommonResponse<Void> update(@RequestBody @Valid Update dto, @CurrentUser Long userId) {
+    public CommonResponse<Void> update(@RequestBody @Valid UserRequestDto.Update dto, @CurrentUser Long userId) {
         userUseCase.update(dto, userId);
         return CommonResponse.createSuccess();
     }
@@ -61,7 +61,7 @@ public class UserController {
     }
 
     @PostMapping("/refresh")
-    public CommonResponse<UserDTO.refreshResponse> refresh(@Valid @RequestBody UserDTO.refreshRequest dto, HttpServletRequest request) {
+    public CommonResponse<UserResponseDto.refreshResponse> refresh(@Valid @RequestBody UserRequestDto.refreshRequest dto, HttpServletRequest request) {
         return CommonResponse.createSuccess(userManageUseCase.refresh(dto, request));
     }
 }
